requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements-base.in requirements-dev.in requirements.in
#
aiofiles==23.2.1
    # via gcloud-aio-storage
aiohttp==3.9.1
    # via gcloud-aio-auth
aiosignal==1.3.1
    # via aiohttp
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   redis
attrs==23.2.0
    # via aiohttp
backoff==2.2.1
    # via gcloud-aio-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via gcloud-aio-auth
click==8.1.7
    # via
    #   -r requirements-base.in
    #   uvicorn
cryptography==42.0.3
    # via gcloud-aio-auth
environs==9.5.0
    # via
    #   -r requirements-base.in
    #   gundi-client-v2
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
fastapi==0.103.2
    # via -r requirements-base.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==5.3.2
    # via
    #   gcloud-aio-pubsub
    #   gcloud-aio-storage
gcloud-aio-pubsub==6.0.0
    # via -r requirements-base.in
gcloud-aio-storage==9.3.0
    # via -r requirements.in
gundi-client-v2==2.3.8
    # via -r requirements-base.in
gundi-core==1.7.1
    # via
    #   -r requirements-base.in
    #   gundi-client-v2
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via
    #   gundi-client-v2
    #   respx
idna==3.6
    # via
    #   anyio
    #   httpx
    #   yarl
iniconfig==2.0.0
    # via pytest
marshmallow==3.20.2
    # via environs
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
packaging==23.2
    # via
    #   marshmallow
    #   pytest
pluggy==1.5.0
    # via pytest
prometheus-client==0.20.0
    # via gcloud-aio-pubsub
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via gcloud-aio-storage
pycparser==2.21
    # via cffi
pydantic==1.10.19
    # via
    #   -r requirements-base.in
    #   fastapi
    #   gundi-client-v2
    #   gundi-core
pyjq==2.6.0
    # via -r requirements-base.in
pyjwt==2.8.0
    # via gcloud-aio-auth
pytest==7.4.4
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==0.21.2
    # via -r requirements-dev.in
pytest-mock==3.12.0
    # via -r requirements-dev.in
python-dotenv==1.0.1
    # via environs
python-json-logger==2.0.7
    # via -r requirements-base.in
redis==5.0.1
    # via -r requirements-base.in
respx==0.20.2
    # via gundi-client-v2
rsa==4.9
    # via gcloud-aio-storage
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
stamina==23.2.0
    # via -r requirements-base.in
starlette==0.27.0
    # via fastapi
tenacity==8.2.3
    # via stamina
tomli==2.2.1
    # via pytest
typing-extensions==4.9.0
    # via
    #   fastapi
    #   pydantic
    #   stamina
    #   starlette
    #   uvicorn
uvicorn==0.23.2
    # via -r requirements-base.in
xmltodict==0.13.0
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
